refactor: split lien_service.go into focused modules
Split the 1322-line lien_service.go into four focused files:
- lien_service.go (64 lines): errors and metric constants
- lien_lifecycle.go (762 lines): InitiateLien, ExecuteLien, TerminateLien
- lien_queries.go (104 lines): RetrieveLien, GetActiveAmountBlocks
- lien_helpers.go (430 lines): proto converters, balance hydration, idempotency checks

All files remain in the same package. No behavior changes.
