+++ services/current-account/service/lien_queries.go
@@ -1,0 +1,104 @@
+package service
+
+import (
+	"context"
+	"errors"
+	"time"
+
+	"github.com/google/uuid"
+	pb "github.com/meridianhub/meridian/api/proto/meridian/current_account/v1"
+	"github.com/meridianhub/meridian/services/current-account/adapters/persistence"
+	caobservability "github.com/meridianhub/meridian/services/current-account/observability"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+)
+
+// RetrieveLien gets lien details
+func (s *Service) RetrieveLien(ctx context.Context, req *pb.RetrieveLienRequest) (*pb.RetrieveLienResponse, error) {
+	start := time.Now()
+	operationStatus := operationStatusSuccess
+	defer func() {
+		caobservability.RecordOperationDuration("retrieve_lien", operationStatus, time.Since(start))
+	}()
+
+	// Validate lien repository is configured
+	if s.lienRepo == nil {
+		operationStatus = opStatusLienRepoNil
+		return nil, status.Error(codes.FailedPrecondition, "lien operations not configured")
+	}
+
+	// Parse lien ID
+	lienID, err := uuid.Parse(req.LienId)
+	if err != nil {
+		operationStatus = opStatusInvalidLienID
+		return nil, status.Errorf(codes.InvalidArgument, "invalid lien ID: %v", err)
+	}
+
+	// Retrieve lien (context is passed for organization scoping in multi-org mode)
+	lien, err := s.lienRepo.FindByID(ctx, lienID)
+	if err != nil {
+		if errors.Is(err, persistence.ErrLienNotFound) {
+			operationStatus = opStatusLienNotFound
+			return nil, status.Errorf(codes.NotFound, "lien not found: %s", req.LienId)
+		}
+		operationStatus = opStatusRetrieveFailed
+		return nil, status.Errorf(codes.Internal, "failed to retrieve lien: %v", err)
+	}
+
+	return &pb.RetrieveLienResponse{
+		Lien: toLienProto(lien),
+	}, nil
+}
+
+// GetActiveAmountBlocks retrieves active fund reservations for Position Keeping.
+// Returns all ACTIVE (non-expired) liens mapped to AmountBlock representation.
+// Used by Position Keeping service to query blocked amounts without coupling to lien details.
+func (s *Service) GetActiveAmountBlocks(ctx context.Context, req *pb.GetActiveAmountBlocksRequest) (*pb.GetActiveAmountBlocksResponse, error) {
+	start := time.Now()
+	operationStatus := operationStatusSuccess
+	defer func() {
+		caobservability.RecordOperationDuration("get_active_amount_blocks", operationStatus, time.Since(start))
+	}()
+
+	// Validate lien repository is configured
+	if s.lienRepo == nil {
+		operationStatus = opStatusLienRepoNil
+		return nil, status.Error(codes.FailedPrecondition, "lien operations not configured")
+	}
+
+	// Retrieve account to validate it exists and get the internal UUID
+	account, err := s.repo.FindByID(ctx, req.AccountId)
+	if err != nil {
+		if errors.Is(err, persistence.ErrAccountNotFound) {
+			operationStatus = opStatusAccountNotFound
+			return nil, status.Errorf(codes.NotFound, "account not found: %s", req.AccountId)
+		}
+		operationStatus = opStatusRetrieveFailed
+		return nil, status.Errorf(codes.Internal, "failed to retrieve account: %v", err)
+	}
+
+	// Retrieve active liens for the account
+	liens, err := s.lienRepo.FindActiveByAccountID(ctx, account.ID())
+	if err != nil {
+		operationStatus = opStatusRetrieveFailed
+		s.logger.Error("failed to retrieve active liens",
+			"account_id", req.AccountId,
+			"error", err)
+		return nil, status.Errorf(codes.Internal, "failed to retrieve active liens: %v", err)
+	}
+
+	// Convert liens to AmountBlock representation
+	blocks := make([]*pb.AmountBlock, 0, len(liens))
+	for _, lien := range liens {
+		block := toAmountBlockProto(lien)
+		blocks = append(blocks, block)
+	}
+
+	s.logger.Debug("retrieved active amount blocks",
+		"account_id", req.AccountId,
+		"block_count", len(blocks))
+
+	return &pb.GetActiveAmountBlocksResponse{
+		Blocks: blocks,
+	}, nil
+}
